Forbid duplicate properties.
This flags a potentially ambiguous form of JSON: objects with
duplicate properties, like

```json
{
  "type": "Feature",
  "type": "FeatureCollection"
}
```

These are valid JSON according to the specification but many
parsers will reject this input, and others will accept the second
value. This PR updates jsonlint to a new version that adds a
non-enumerable property that keeps track of duplicate properties (1.7.1)
and then adds a default-true option to forbid these properties.